fix(input): taro demo拆解&文档引入
