+++ src/packages/input/demos/taro/demo1.tsx
@@ -4,7 +4,12 @@ import { Input } from '@nutui/nutui-react-taro'
 const Demo1 = () => {
   return (
     <>
-      <Input placeholder="请输入文本" />
+      <Input
+        placeholder="请输入文本"
+        onChange={(v) => {
+          console.log('onChange', v)
+        }}
+      />
     </>
   )
 }
